Refactored Im2Col tests - now taking advantage of templates to eliminate code duplication
